chore: restructure components directories
