+++ src/components/layout/Header.tsx
@@ -1,9 +1,9 @@
 import { LogoLink } from '@/components/layout/LogoLink';
 import { MobileSheet } from '@/components/layout/MobileSheet';
 import { Nav } from '@/components/layout/Nav';
-import { DarkModeMenu } from '@/components/settings/DarkModeMenu';
-import { LocaleMenu } from '@/components/settings/LocaleMenu';
-import { ProfileMenu } from '@/components/settings/ProfileMenu';
+import { DarkModeMenu } from '@/components/layout/header/DarkModeMenu';
+import { LocaleMenu } from '@/components/layout/header/LocaleMenu';
+import { ProfileMenu } from '@/components/layout/header/ProfileMenu';
 import { useTranslations } from 'next-intl';
 
 function Header() {
